test latency

+++ core/models.py
@@ -15,7 +15,7 @@ class Business(models.Model):
     name = models.CharField(max_length=100)
     iae_code = models.CharField(max_length=50, default='', blank=True)
     rentability = models.IntegerField(default=0, validators=[MinValueValidator(0), MaxValueValidator(100)])
-    tipology = models.DecimalField(default=0, validators=[MinValueValidator(0), MaxValueValidator(1)], max_digits=2, decimal_places=1)
+    tipology = models.FloatField(default=0, validators=[MinValueValidator(0), MaxValueValidator(1)])
     proximity_to_urban_center_m = models.IntegerField(default=0)
     coordinates = models.JSONField(default=default_coordinates)
